requirements.txt
flask
supabase
python-dotenv
gunicorn
flask-socketio
redis
eventlet
